- LED CUBE Built / Debugging
- Cube_Effect.cpp - Fixed the Effect_TopDown() and the Effect_Rain().
- Cube_Main.cpp - fixed the timer values as the timer is using a 1024 clk divider and some where based in the old 64 clk divider.
- TODO: The update rate does not appear to match the calculated values, check this

+++ src/Cube_Effect.h
@@ -9,8 +9,8 @@ typedef enum {
 	}EffectEnum;
 
 void	EffectLoop(int Effect);
-void	Effect_Rain(int iterations, int itterationDelay);
-void	Effect_TopDown(byte axis, int seperation, bool direction, int iterations, int itterationDelay);
+void	Effect_Rain(int iterations, unsigned long itterationDelay);
+void	Effect_TopDown(byte axis, int seperation, bool direction, int iterations, unsigned long itterationDelay);
 void	EffectShootText(byte plane, String inputStr, int iterations, int delayTime, int shiftDelayTime);
 void	EffectFireworks(int iterations, int n, int delayTime);
 void	EffectScrollText(int iterations, String inputstr, int delayTime);
